requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
diagrams==0.24.4
graphviz>=0.20.2
Deprecated==1.2.18
matplotlib==3.10.5
